pyproject.toml
[project]
name = "mittation"
version = "0.1.0"
description = "generative ai data migration"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [

    "langgraph (>=0.6.2,<0.7.0)",
    "langchain-google-genai (>=2.1.8,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "google-cloud-aiplatform (>=1.106.0,<2.0.0)",
    "networkx (>=3.5,<4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "graphviz (>=0.21,<0.22)",
    "ipython (>=9.4.0,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
